requirements.txt
beartype
ifstools
Pillow
tqdm
